move KeyRegistries into config, with NetworkConfig
